Move std.nonemptylist functions into each typeclass companion object with JvmStatic annotation and remove imports from deprecations
