+++ arrow-libs/optics/arrow-optics/src/test/kotlin/arrow/optics/instances/NonEmptyListInstanceTest.kt
@@ -8,7 +8,6 @@ import arrow.core.test.UnitSpec
 import arrow.core.test.generators.functionAToB
 import arrow.core.test.generators.nonEmptyList
 import arrow.optics.Traversal
-import arrow.optics.nonEmptyList
 import arrow.optics.test.laws.OptionalLaws
 import arrow.optics.test.laws.TraversalLaws
 import arrow.optics.typeclasses.FilterIndex
